% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{calendar-names}
\alias{calendar-names}
\alias{month_names}
\alias{weekday_names}
\alias{ampm_indicators}
\title{Calendar Names}
\usage{
month_names(locale = NULL, abbreviate = TRUE)

weekday_names(locale = NULL, abbreviate = TRUE)

ampm_indicators(locale = NULL)
}
\arguments{
\item{locale}{a character value determining locale or \code{NULL} (default,
interpreted as the current system locale).}

\item{abbreviate}{a logical value, if \code{TRUE}, abbreviated names are returned;
if \code{FALSE}, full names are returned.
\code{TRUE} by default.}
}
\value{
A character vector of length 12, 7, or 2.
}
\description{
These three functions return (abbreviated and full) names of months and days of week
as well as AM/PM indicators in the current or user-provided locale.

How the month and weekday names are actually returned depends both on the
selected locale and character set / code page setting.
}
\section{Locale Settings}{

Unfortunately, locale and character set naming is not standardised across
different operating systems. Under Linux, Mac, and other Unixes
locale is usually of the form \code{xx_XX} (\code{xx} for language,
\code{XX} for country) optionally followed by a dot and a character set
identifier, for example, \code{UTF-8}. Under Windows locale should be of the form
\code{Language} or \code{Language_Country} optionally followed by a dot
and a 3-5 digit code page identifier. See Examples for the most popular
operating systems.

\code{"C"} is a special locale that should always be available and
defaults to American English.
}

\examples{
\dontrun{
## Linux, Unix, Mac
# English abbreviated month names
month_names("en_GB")
# French month names
month_names("fr_FR.UTF-8", FALSE)
# German abbreviated month names
month_names("de_DE.UTF-8")
# Polish abbreviated month names
month_names("pl_PL.UTF-8")
# English weekday names
weekday_names("en_GB", FALSE)
# French abbreviated weekday names
weekday_names("fr_FR.UTF-8")
# German weekday names
weekday_names("de_DE.UTF-8", FALSE)
# Polish abbreviated weekday names
weekday_names("pl_PL.UTF-8")
# US am/pm indicators
ampm_indicators("en_US")
# UK am/pm indicators
ampm_indicators("en_GB")

## Windows
# English abbreviated month names
month_names("English")
# French month names
month_names("French", FALSE)
# German abbreviated month names
month_names("German")
# Polish abbreviated month names
month_names("Polish")
# English weekday names
weekday_names("English", FALSE)
# French abbreviated weekday names
weekday_names("French")
# German weekday names
weekday_names("German", FALSE)
# Polish abbreviated weekday names
weekday_names("Polish")
# US am/pm indicators
ampm_indicators("English_United States")
# UK am/pm indicators
ampm_indicators("English_United Kingdom")
}

}
\seealso{
\code{\link{format}} for formatting objects of \code{tind} class.
}
